Refactor: Restructure schedule and tasks tree modules for improved organization and clarity

+++ src/domain/Schedule.fs
@@ -10,5 +10,5 @@ type Schedule = {
     StopTime: TimeOnly option
     Workdays: DayOfWeek Set
     Recursively: bool
-    TimeZone: float
+    TimeZone: byte
 }
